// src/constant/types.ts
export interface News {
  author: string[];
  link: string;
  photo_url: string;
  published_datetime_utc: string;
  snippet: string;
  source_favicon_url: string;
  source_logo_url: string;
  source_name: string;
  source_publication_id: string;
  source_url: string;
  story_id: string;
  sub_articles: string[];
  thumbnail_url: string;
  title: string;
}